Create Report-file-data.txt
This text file is the data file you need to use in your code. (This is the actual "report" that they gave you to work on.
In your code you need to import this file, and then do your "coding" part to get the answers they need to complete the challenge.